+++ src/main/java/com/example/SEENEMA/post/view/service/ViewPostServiceImpl.java
@@ -1,6 +1,7 @@
 package com.example.SEENEMA.post.view.service;
 
 import com.example.SEENEMA.post.file.Image;
+import com.example.SEENEMA.post.view.dto.ResponseMessage;
 import com.example.SEENEMA.post.view.dto.ViewPostDto;
 import com.example.SEENEMA.post.file.ImageRepository;
 import com.example.SEENEMA.post.view.repository.ViewPostRepository;
@@ -17,6 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -55,7 +57,7 @@ public class ViewPostServiceImpl implements ViewPostService {
 
     @Override
     @Transactional(readOnly = true)
-    public ViewPostDto.detailResponse readViewPost(Long userId, Long theaterId, Long viewNo){
+    public ViewPostDto.detailResponse readViewPost(Long theaterId, Long viewNo){
 
         ViewPost viewPost = getViewPost(theaterId,viewNo);
         // 이미지 컬렉션을 명시적으로 초기화
@@ -65,20 +67,41 @@ public class ViewPostServiceImpl implements ViewPostService {
 
     @Override
     @Transactional
-    public ViewPostDto.addResponse updateViewPost(Long theaterId,Long viewNo, ViewPostDto.updateRequest requestDto){
-
-        ViewPost viewPost = getViewPost(theaterId,viewNo);
-        viewPost.updateViewPost(requestDto.getPlay(), requestDto.getSeat(), requestDto.getTitle(), requestDto.getContent(), requestDto.getImage());
-
-        return new ViewPostDto.addResponse(viewPost);
+    public String authUserForEdit(Long theaterId, Long viewNo, Long userId){
+        ViewPost viewPost = getViewPost(theaterId, viewNo);
+        if(viewPost.getUser().getUserId().equals(userId)) return "SUCCESS";
+        else return "NOT_SAME_USER";
     }
 
     @Override
     @Transactional
-    public void deleteViewPost(Long theaterId, Long viewNo){
+    public ViewPostDto.addResponse updateViewPost(Long theaterId,Long viewNo, ViewPostDto.updateRequest requestDto, Long userId){
 
+        ViewPost viewPost = getViewPost(theaterId,viewNo);
+        // update전 작성자와 사용자 동일인 판별
+        if(!viewPost.getUser().getUserId().equals(userId)) {
+            // 동일인 X -> 수정 X
+            return new ViewPostDto.addResponse(viewPost);
+        }
+        else{
+            viewPost.updateViewPost(requestDto.getPlay(), requestDto.getSeat(), requestDto.getTitle(), requestDto.getContent(), requestDto.getImage());
+
+            return new ViewPostDto.addResponse(viewPost);
+        }
+    }
+
+    @Override
+    @Transactional
+    public String deleteViewPost(Long theaterId, Long viewNo, Long userId){
+        // 시야 후기 게시글 삭제
         ViewPost viewPost = getViewPost(theaterId, viewNo);
-        viewPostRepository.delete(viewPost);
+        if(viewPost.getUser().getUserId().equals(userId)) {
+            viewPostRepository.delete(viewPost);
+            return ResponseMessage.DELETE.getMsg();
+        }
+        else {
+            return "FAIL";
+        }
     }
 
     @Override
